Fix not able to navigate map if mouse is inside minimap.
On laptops, for example, it's nice to be able to click in the minimap to navigate around, but then you often want to quickly navigate via keyboard. The current behavior is that if you click in the minimap, then the keyboard does not work, which doesn't make sense since the minimap doesn't do anything with these events anyway. So, we can pass the minimap events onto the rest of the game map. This way you can click in the minimap and then use arrow keys to move. We simply ignore mouse events when in minimap so that the cursor does not update based on map content behind the minimap.