Cache Rendered Fonts (#94)
* Cache Rendered Fonts

This puts the font rendering in a pool of surfaces, cached by text and color. This is an optimization to prevent random allocations per-frame. It periodically garbage collects, on a safe interval to prevent memory leaks. Based on testing this does not cause any dropped frames with the current number of rendered fonts.

* Don't iterate map every tick + add debug logs

* move const to top of file