maps: Add Arena Fut
This is from the user maps file provided by @ph2708 in
https://github.com/netpanzer/netpanzer/discussions/76

I don't know who made the map.

I'm not familiar enough with NetPanzer yet to determine if a map is
"good" or "playable", so this will definitely need a good review. I was
able to connect to a server hosting this map with no warnings or errors.
The map looks good to me.